Utilising the Vuex Store
Another great feature of Nuxt.js is that it supports Vuex out of the box. In this lesson, we're learning what we need to know to work with Vuex and Nuxt: https://vuex.vuejs.org/

We will not cover what Vuex is in this lesson. Check out our Vuex for Everyone course if you want to learn Vuex: https://vueschool.io/courses/vuex-for-everyone

https://vueschool.io/lessons/utilising-the-vuex-store-nuxtjs